+++ sdk/core/test/integration/otp/core.test.ts
@@ -126,8 +126,13 @@ parallel('CommonNetworkMember [OTP]', () => {
 
     await commonNetworkMember.completeChangeEmailOrPhone(changeToken, changeUsernameCode)
     await commonNetworkMember.logOut()
+    // NOTE: try/catch added as a workaround because of issue NotAuthorizedException see https://github.com/aws-amplify/amplify-js/issues/9838
+    try {
+      commonNetworkMember = await AffinidiWallet.logInWithPassword(options, newInbox.email, password)
+    } catch (err) {
+      commonNetworkMember = await AffinidiWallet.logInWithPassword(options, newInbox.email, password)
+    }
 
-    commonNetworkMember = await AffinidiWallet.logInWithPassword(options, newInbox.email, password)
     checkIsWallet(commonNetworkMember)
 
     await commonNetworkMember.logOut()
@@ -170,8 +175,13 @@ parallel('CommonNetworkMember [OTP]', () => {
 
     await commonNetworkMember.completeChangeEmailOrPhone(changeToken, changeUsernameCode)
     await commonNetworkMember.logOut()
+    // NOTE: try/catch added as a workaround because of issue NotAuthorizedException see https://github.com/aws-amplify/amplify-js/issues/9838
+    try {
+      commonNetworkMember = await AffinidiWallet.logInWithPassword(options, newInbox.email, password)
+    } catch (err) {
+      commonNetworkMember = await AffinidiWallet.logInWithPassword(options, newInbox.email, password)
+    }
 
-    commonNetworkMember = await AffinidiWallet.logInWithPassword(options, newInbox.email, password)
     checkIsWallet(commonNetworkMember)
   })