+++ common-libs/user-management/CHANGELOG.md
@@ -1,3 +1,5 @@
+# release 1.4.0 (2022-06-16)
+* add logInWithRefreshToken auth method
 # release 1.3.1 (2022-06-22)
 * updated `internal-api-clients`, `tools-common` with BigInt fix for react native
 # release 1.3.0 (2022-06-10)