+++ sdk/node/CHANGELOG.md
@@ -1,3 +1,5 @@
+# release 6.3.0 (2022-06-29)
+* add login with refreshToken
 # release 6.2.9 (2022-06-28)
 * fix bug in a fetch credential logic
 # release 6.2.8 (2022-06-28)
@@ -6,7 +8,7 @@
 - `userPoolId` & `clientId` configure custom cognito pool & client. use environment specific settings by default
 # release 6.2.7 (2022-06-22)
 * updated `wallet-core-sdk`, `common` with BigInt fix for react native
-* # release 6.2.6 (2022-06-10)
+# release 6.2.6 (2022-06-10)
 * fix handling of user registration status incomplete
 # release 6.2.5 (2022-06-06)
 * update README to notify users use the latest versions with affinidi-vault support